Update setup.py
author: Specifies the name of the package author.
author_email: Specifies the email address of the package author.
install_requires: Specifies the dependencies required by the package. Add the necessary dependencies as a list of strings.
entry_points: Allows you to define command-line scripts or other entry points for your package. It associates the sgm-cli command with the main() function in the sgm.cli module.
package_data: Specifies additional files or data to be included with the package. In this example, it includes all text files in the data directory of the sgm package.
license: Specifies the license under which the package is distributed (e.g., "MIT", "Apache License 2.0", etc.).
classifiers: Provides classifiers or tags for the package that indicate its properties, target audience, and compatibility.
Feel free to modify these fields according to your package's specific requirements.