Update discretizer.py
Using CPU as the default device. 

The models are initialized way before the `init_process_group()` is invoked. The Discretizer is placed in the GPU, and at this stage it is always the first GPU in the node. This initializes some CUDA buffers there that are not liberated when the model is placed in the right place. Initializing the `Discretizer` in the CPU fixes the problem.