improve test_inference.py
I guess, we can move the creation of `SamplingPipeline` objects and clearing the CUDA cache outside the test fixtures and execute them only once before and after the entire test class. This way, we reduce the overhead of object creation and deletion for each test case. Is something like this right?